Split project/project.go into smaller files
